Happy with client.py, formatting matches assignment
